third_party/WebKit/LayoutTests/http/tests/security/referrer-on-client-reload.html

<!DOCTYPE html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body></body>

<script>
async_test(function () {
  var test = this;
  var saw_original_referrer = false;
  window.addEventListener("message", test.step_func(function (event) {
    assert_equals(event.data.referrer, "");
    test.done();
  }));

  var iframe = document.createElement("iframe");
  iframe.src = "resources/post-referrer-on-reload.html";
  document.body.appendChild(iframe);
}, "Reloading a document uses the document's referrer policy");
</script>

</html>
